requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1            
python-dotenv==1.1.0
langchain-pinecone==0.2.8
transformers==4.48.1
langchain-community==0.3.26 
huggingface-hub==0.34.4
langchain-huggingface==0.3.1
Markdown==3.9
pinecone==7.3.0
